Изменил 9 строку(price) теперь понимает не только целые числа

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, Numeric
 from .database import Base
 class Item(Base):
     __tablename__ = "items"
@@ -6,4 +6,4 @@ class Item(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(100), nullable=False)
     description = Column(String(225), nullable=True)
-    price = Column(Integer, nullable=False)
+    price = Column(Numeric(12,4), nullable=False)
